--- src/create-i18n/index.ts
@@ -1,1 +1,0 @@
-export * from "./create-i18n";
