+++ src/index.ts
@@ -1,3 +1,7 @@
-export * from "@/create-i18n/index";
-export * from "@/translate/index";
-export * from "@/use-i18n/index";
+import { createI18n } from "@/create-i18n/create-i18n";
+import { t } from "@/translate/translate";
+import { useI18n } from "@/use-i18n/use-i18n";
+import type { Locale } from "@/locale/locale";
+
+export { createI18n, t, useI18n };
+export type { Locale };
