--- src/use-i18n/index.ts
@@ -1,1 +1,0 @@
-export * from "./use-i18n";
